Add files via upload
This project is an 'Online Book Order System' that includes both a user side application and an admin panel. The user can search books and order books, while the admin can can add and delete books and also has access to order details and customer details.
We have used  Streamlit  to create interactive web pages and MySQL database to maintain the data and perform CRUD operation on the data. The 'mysql.connector' library is used to interact with the database in python.